pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rastir"
version = "0.1.0"
description = "Rastir — LLM & Agent Observability Library with decorators, metrics, and traces"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Rastir Contributors" },
]
keywords = ["llm", "observability", "tracing", "metrics", "prometheus", "opentelemetry"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "httpx>=0.25.0",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
server = [
    "rastir[otel]",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "prometheus-client>=0.19.0",
    "pyyaml>=6.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "httpx",  # for test client
]
all = ["rastir[server,dev]"]

[project.scripts]
rastir-server = "rastir.server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rastir"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
